requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
easyocr==1.7.2
et_xmlfile==2.0.0
filelock==3.25.1
Flask==3.1.3
fsspec==2026.2.0
ImageIO==2.37.3
itsdangerous==2.2.0
Jinja2==3.1.6
lazy-loader==0.5
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.4.2
ninja==1.13.0
numpy==2.2.6
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
openpyxl==3.1.5
packaging==26.0
pandas==2.3.3
pillow==12.1.1
pyclipper==1.4.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
pyngrok==7.1.6
pytz==2026.1.post1
PyYAML==6.0.3
scikit-image==0.25.2
scipy==1.15.3
shapely==2.1.2
six==1.17.0
sympy==1.14.0
tifffile==2025.5.10
torch==2.10.0
torchaudio==2.10.0
torchvision==0.25.0
typing_extensions==4.15.0
tzdata==2025.3
Werkzeug==3.1.6
PyMuPDF==1.25.5
